ID type changed from int to long

+++ src/main/java/com/example/travel_buddy_app/repositories/HostRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.Query;
 
 import java.util.Optional;
 
-public interface HostRepository extends JpaRepository<Host,Integer> {
+public interface HostRepository extends JpaRepository<Host,Long> {
     @Query(name = "findHostByID", value = "select t from Host t where t.id=?1")
     public Optional<Host> findHostEntityByID(int id);
 }
